refactor(db): split migration.rs into module with subfiles
Split migration.rs (3071 lines) into 4 files:
- migration/mod.rs (177 lines): trait, all_migrations(), shared helpers
- migration/v1.rs (731 lines): V1InitialSchema
- migration/v2_v5.rs (1387 lines): V2-V5 migrations
- migration/v41_v46.rs (689 lines): V41-V46 migrations

957 tests pass.
